Completed building the cart and wishlist layers in the backend

+++ quickbuy/quickbuy/src/main/java/com/ecommerce/quickbuy/repository/CartRepository.java
@@ -9,6 +9,6 @@ import java.util.Optional;
 @Repository
 public interface CartRepository extends JpaRepository<Cart, Integer> {
 
-    Optional<Cart> findByUserId(Long userId);
+    Optional<Cart> findByUserId(int userId);
 
 }
